fix: improve tests

+++ server/internal/database/migrations/00002_account_table.sql
@@ -4,7 +4,7 @@
 CREATE TABLE IF NOT EXISTS ${DB_SCHEMA}.account (
     id SERIAL PRIMARY KEY,
     name VARCHAR(100) NOT NULL,
-    balance FLOAT NOT NULL DEFAULT 0,
+    balance DECIMAL(10, 2) NOT NULL DEFAULT 0,
     bank_type VARCHAR(100) NOT NULL,
     currency VARCHAR(10) NOT NULL,
     created_by INTEGER NOT NULL,
